Use layout/page for Functions mimicking the structure for Streams
Without this, having a SlideOver visible would not show the background. Not sure why

https://github.com/inngest/inngest/blob/5083de28998acc02916baba7b6369d246dc7edd3/ui/apps/dev-server-ui/src/app/(dashboard)/stream/layout.tsx
https://github.com/inngest/inngest/blob/5083de28998acc02916baba7b6369d246dc7edd3/ui/apps/dev-server-ui/src/app/(dashboard)/stream/page.tsx
