+++ ui/apps/dev-server-ui/src/app/(dashboard)/functions/layout.tsx
@@ -1,0 +1,14 @@
+import FunctionList from '@/app/(dashboard)/functions/FunctionList';
+
+type FunctionLayoutProps = {
+  children: React.ReactNode;
+};
+
+export default function FunctionLayout({ children }: FunctionLayoutProps) {
+  return (
+    <>
+      <FunctionList />
+      {children}
+    </>
+  );
+}
